v1.0.0.0 (built): web hardening, content loaders, UI scaffolding, build log
C#: Implemented robust FileContentLoader (text/JSON/XML/HTML/images/binaries), encoding detection, safe HTML preview, DI registration.

C#: WebContentLoader: removed System.Web.HttpUtility; added custom query parser; expanded curated sources (NASA, USGS, Open‑Meteo, Wikimedia/Wikipedia, OpenStreetMap tiles, YouTube‑nocookie); safer iframe sandbox & referrerpolicy.

C#: WebViewRenderer: expanded allowlist; fixed localhost handling; removed unsupported CoreWebView2.Stop(); removed placeholder/redacted host checks. app.manifest version corrected to 1.0.0.0.

C++ UI: Added Dashboard tab (CPU/Memory/Network tiles) and Task Manager scaffolding; Options window scaffolding for tab wiring.

Build: RenderProcess builds successfully (Debug); logged warnings (System.Text.Json vulnerability advisories; OpenTK TFM compatibility).

+++ renderprocess/Communication/IPCMessageHandler.cs
@@ -512,7 +512,7 @@ public class IPCMessageHandler : IDisposable
             byte[] commandData = new byte[dataSize];
             bytesRead = await _namedPipe.ReadAsync(commandData, 0, (int)dataSize);
 
-            if (bytesRead != dataSize)
+            if (bytesRead != (int)dataSize)
             {
                 _logger.LogError($"Incomplete command read: expected {dataSize}, got {bytesRead}");
                 return false;
@@ -647,7 +647,7 @@ public class IPCMessageHandler : IDisposable
                 WidgetId = command.WidgetId,
                 Status = RenderResultStatus.Failure,
                 ErrorMessage = $"Command handling failed: {ex.Message}",
-                Timestamp = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds()
+                Timestamp = (ulong)DateTimeOffset.UtcNow.ToUnixTimeMilliseconds()
             };
 
             await SendResultAsync(errorResult);
